Fix issue with duplicated Chicago bills. (#34)
This fixes Issue #31 by updating the Chicago bill scraper.

The join of bills with actions is replaced with a subquery that selects
all relevant bill action history into a JSON string, allowing all bill
information to be represented in a single row. This is probably less
performant than the join was, but at the scale of these queries it
shouldn't matter too much.

This changes the structure of internal types, but the schema of data
returned by the scraper is not changed.